% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download.CSV}
\alias{download.CSV}
\title{Downloads CSV}
\usage{
\method{download}{CSV}(x, message.on.fail = TRUE, ..., sep = ";")
}
\arguments{
\item{x}{url of csv file location}

\item{message.on.fail}{logical: show message on download failure?}

\item{...}{optional additional arguments to download function}

\item{sep}{field separator}
}
\description{
Downloads CSV
}

